+++ CHANGELOG.md
@@ -6,12 +6,29 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 Currently, this repo is in Prerelease. When it is released, this project will adhere to [Semantic Versioning](https://semver.org/spec/v2.0.0.html)
 ========
 
+## Prerelease
+
+### Breaking Changes
+
+-   `FormDropDown` and `Dropdown` are now `Select`
+-   Changes `referenceId` to `htmlFor` on `Label`
+
+### Adds
+
+-   `aria-live` and `aria-atomic` to `HelperErrorText` in its errored state
+-   Adds "Optional", "Required", or empty string inside `Label` for forms
+
+### Changes
+
+-   Fixes bug in former `FormDropdown` (now `Select`) where the select value would not change if `selectedOption` were passed
+-   `Link`'s scss now references `$ui-link-primary` instead of `$nypl-blue-regular`
+-   Changes `Select`'s SCSS to use the updated styling from Filament Group
+
 ## 0.6.0
 
 ### Breaking Changes
 
 -   Removes `RNSectionTitle`
--   `FormDropDown` and `Dropdown` are now `Select`
 
 ### Changes